This is a slightly Different Approach to processing with an Analyzer
Checking in this code - But this is a breaking change. The analyzers are running in a Foreach Loop, the end-result is it creates some crazy compilation, which pretty much kills your machine. The focus here, is the start/stop Compilation for the Analyzer, which is why this code is added. It's analysis only, and should be changed.